updating

+++ v1/app/services/facade.py
@@ -118,7 +118,8 @@ class HBnBFacade:
         check_id = self.amenity_repo.get(amenity_id)
         if check_id is None:
             raise ValueError (f"Amenity with ID {amenity_id} not found")
-        return self.amenity_repo.update(amenity_id, amenity_data)
+        self.amenity_repo.update(amenity_id, amenity_data)
+        return self.amenity_repo.get(amenity_id)
 
     # --- CRU Review ---
     def create_review(self, review_data):
@@ -179,4 +180,8 @@ class HBnBFacade:
         return review
 
     def delete_review(self, review_id):
-        return self.review_repo.delete(review_id)
+        review = self.review_repo.get(review_id)
+        if review:
+            self.review_repo.delete(review_id)
+            return True
+        return False
